+++ packages/lumx-react/src/components/image-block/ImageBlock.tsx
@@ -4,7 +4,7 @@ import classNames from 'classnames';
 
 import isObject from 'lodash/isObject';
 
-import { Alignment, AspectRatio, FocusPoint, Size, Theme, Thumbnail } from '@lumx/react';
+import { Alignment, AspectRatio, CrossOrigin, FocusPoint, Size, Theme, Thumbnail } from '@lumx/react';
 
 import { COMPONENT_PREFIX } from '@lumx/react/constants';
 import { GenericProps, getRootClassName, handleBasicClasses } from '@lumx/react/utils';
@@ -34,6 +34,11 @@ interface ImageBlockProps extends GenericProps {
     captionPosition?: ImageBlockCaptionPosition;
     /** The style to apply to the caption section. */
     captionStyle?: CSSProperties;
+    /**
+     * Allows images that are loaded from foreign origins
+     * to be used as if they had been loaded from the current origin.
+     */
+    crossOrigin?: CrossOrigin;
     /** The image description. Can be either a string, or sanitized html. */
     description?:
         | string
@@ -46,6 +51,12 @@ interface ImageBlockProps extends GenericProps {
     focusPoint?: FocusPoint;
     /** The url of the image we want to display in the image-block. */
     image: string;
+    /** Active cross origin. */
+    isCrossOriginEnabled?: boolean;
+    /** Allows to re-center the image according to the focal point after after window resizing */
+    isFollowingWindowSize?: boolean;
+    /** Time before recalculating focal point if isFollowingWindowSize is activated */
+    resizeDebounceTime?: number;
     /** The image block size. */
     size?: ImageBlockSize;
     /** Tags elements to be transcluded into the component */
@@ -80,9 +91,13 @@ const DEFAULT_PROPS: DefaultPropsType = {
     aspectRatio: AspectRatio.original,
     captionPosition: ImageBlockCaptionPosition.below,
     captionStyle: {},
+    crossOrigin: undefined,
     description: undefined,
     fillHeight: false,
     focusPoint: undefined,
+    isCrossOriginEnabled: undefined,
+    isFollowingWindowSize: undefined,
+    resizeDebounceTime: undefined,
     size: undefined,
     tags: undefined,
     theme: Theme.light,
@@ -101,10 +116,14 @@ const ImageBlock: React.FC<ImageBlockProps> = ({
     className = '',
     captionPosition = DEFAULT_PROPS.captionPosition,
     captionStyle = DEFAULT_PROPS.captionStyle,
+    crossOrigin = DEFAULT_PROPS.crossOrigin,
     description = DEFAULT_PROPS.description,
     fillHeight = DEFAULT_PROPS.fillHeight,
     focusPoint = DEFAULT_PROPS.focusPoint,
     image,
+    isCrossOriginEnabled = DEFAULT_PROPS.isCrossOriginEnabled,
+    isFollowingWindowSize = DEFAULT_PROPS.isFollowingWindowSize,
+    resizeDebounceTime = DEFAULT_PROPS.resizeDebounceTime,
     size = DEFAULT_PROPS.size,
     tags = DEFAULT_PROPS.tags,
     theme = DEFAULT_PROPS.theme,
@@ -137,11 +156,15 @@ const ImageBlock: React.FC<ImageBlockProps> = ({
                 align={align}
                 className={`${CLASSNAME}__image`}
                 aspectRatio={aspectRatio}
+                crossOrigin={crossOrigin}
+                isCrossOriginEnabled={isCrossOriginEnabled}
+                isFollowingWindowSize={isFollowingWindowSize}
                 size={size}
                 fillHeight={fillHeight}
                 focusPoint={focusPoint}
                 image={image}
                 onClick={onClick}
+                resizeDebounceTime={resizeDebounceTime}
                 theme={theme}
             />
             {(title || description || tags) && (